requirements.txt
absl-py==2.1.0
ale-py==0.8.1
box2d-py==2.3.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.37
decorator==4.4.2
etils==1.8.0
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
glfw==2.7.0
gymnasium==0.29.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib_resources==6.4.0
jax==0.4.28
jaxlib==0.4.28
Jinja2==3.1.4
kiwisolver==1.4.5
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib==3.9.0
ml-dtypes==0.4.0
moviepy==1.0.3
mpmath==1.3.0
mujoco==3.1.5
mujoco-py==2.1.2.14
neat-python==0.92
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pillow==10.3.0
proglog==0.1.10
pycparser==2.22
pygame==2.5.2
PyOpenGL==3.1.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.2
scipy==1.13.1
Shimmy==0.2.1
six==1.16.0
swig==4.2.1
sympy==1.12
torch==2.3.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
zipp==3.18.2
